+++ spec/create-spec.js
@@ -497,10 +497,10 @@ describe('create', () => {
 		});
 	});
 	describe('runtime support', () => {
-		it('creates node 6.10 deployments by default', done => {
+		it('creates node 8.10 deployments by default', done => {
 			createFromDir('hello-world')
 			.then(getLambdaConfiguration)
-			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs6.10'))
+			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs8.10'))
 			.then(done, done.fail);
 		});
 		it('can create nodejs4.3 when requested', done => {
@@ -508,6 +508,13 @@ describe('create', () => {
 			createFromDir('hello-world')
 			.then(getLambdaConfiguration)
 			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs4.3'))
+			.then(done, done.fail);
+		});
+		it('can create nodejs6.10 when requested', done => {
+			config.runtime = 'nodejs6.10';
+			createFromDir('hello-world')
+			.then(getLambdaConfiguration)
+			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs6.10'))
 			.then(done, done.fail);
 		});
 	});
@@ -634,7 +641,7 @@ describe('create', () => {
 				});
 			})
 			.then(() => lambda.getFunctionConfiguration({ FunctionName: 'hello-world2' }).promise())
-			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs6.10'))
+			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs8.10'))
 			.then(done, done.fail);
 		});
 		it('renames scoped NPM packages to a sanitized Lambda name', done => {
@@ -648,7 +655,7 @@ describe('create', () => {
 				});
 			})
 			.then(() => lambda.getFunctionConfiguration({ FunctionName: 'test_hello-world' }).promise())
-			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs6.10'))
+			.then(lambdaResult => expect(lambdaResult.Runtime).toEqual('nodejs8.10'))
 			.then(done, done.fail);
 		});
 		it('uses the package.json description field if --description is not provided', done => {